chatbot/requirements.txt
# WealthArena Runtime Dependencies
# Canonical runtime requirements used by scripts and Docker
# This file contains all dependencies needed to run the API and ingestion scripts

# Azure Deployment Notes
# - chromadb (line 19) has heavy dependencies with native extensions
# - First Azure deployment may take 15-20 minutes due to compilation
# - The .deployment file and SCM_BUILD_TIMEOUT app setting accommodate this
# - Subsequent deployments are faster due to Azure's build caching

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
httpx>=0.27.0
pydantic==2.5.0
yfinance==0.2.18
# Updated for better compatibility
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0
prometheus-client==0.19.0
websockets==12.0

# Vector database for Knowledge Base
# Heavy dependency - requires extended build time on Azure (10-15 min for first deployment)
# Using compatible version that works with pandas/numpy
chromadb==0.4.22
posthog==2.4.2

# Async HTTP client
aiohttp==3.12.15

# Document processing
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
